Moved randomVariableFromScope to GenerateArbitraryExpression
